+++ hmac-mapper/src/main/java/de/intension/protocol/oidc/mappers/HmacPairwiseEmailMapper.java
@@ -3,11 +3,7 @@ package de.intension.protocol.oidc.mappers;
 import java.util.List;
 
 import org.keycloak.common.util.ObjectUtil;
-import org.keycloak.models.ClientSessionContext;
-import org.keycloak.models.KeycloakSession;
-import org.keycloak.models.ProtocolMapperModel;
-import org.keycloak.models.UserModel;
-import org.keycloak.models.UserSessionModel;
+import org.keycloak.models.*;
 import org.keycloak.protocol.oidc.mappers.OIDCAttributeMapperHelper;
 import org.keycloak.provider.ProviderConfigProperty;
 import org.keycloak.representations.AccessToken;
@@ -25,41 +21,46 @@ import org.keycloak.utils.StringUtil;
  * @see <a href=
  *      "https://docs.oracle.com/en/java/javase/11/docs/specs/security/standard-names.html#mac-algorithms">mac-algorithms</a>
  */
-public class HmacPairwiseEmailMapper extends HmacPairwiseSubMapper {
+public class HmacPairwiseEmailMapper extends HmacPairwiseSubMapper
+{
 
-    public static final String PROVIDER_ID = "oidc-hmac-pairwise-email-mapper";
+    public static final String  PROVIDER_ID             = "oidc-hmac-pairwise-email-mapper";
 
-    private static final String EMAIL_DOMAIN_PROP_NAME = "emailDomain";
+    private static final String EMAIL_DOMAIN_PROP_NAME  = "emailDomain";
     private static final String EMAIL_DOMAIN_PROP_LABEL = "Email domain";
-    private static final String EMAIL_DOMAIN_PROP_HELP = "The email domain is appended to the claim value after pseudonymization.";
+    private static final String EMAIL_DOMAIN_PROP_HELP  = "The email domain is appended to the claim value after pseudonymization.";
 
-    private static final String OVERRIDE_PROP_NAME = "override";
-    private static final String OVERRIDE_PROP_LABEL = "Override existing";
-    private static final String OVERRIDE_PROP_HELP = "Toggle overriding an existing email claim.";
+    private static final String OVERRIDE_PROP_NAME      = "override";
+    private static final String OVERRIDE_PROP_LABEL     = "Override existing";
+    private static final String OVERRIDE_PROP_HELP      = "Toggle overriding an existing email claim.";
 
     @Override
-    public String getId() {
+    public String getId()
+    {
         return PROVIDER_ID;
     }
 
     @Override
-    public String getDisplayType() {
+    public String getDisplayType()
+    {
         return "HMAC pairwise email";
     }
 
     @Override
-    public String getHelpText() {
+    public String getHelpText()
+    {
         return "Updates the email claim in tokens with a pseudonymized value from a selected attribute.";
     }
 
     @Override
     public IDToken transformIDToken(IDToken token, ProtocolMapperModel mappingModel, KeycloakSession session,
-            UserSessionModel userSession, ClientSessionContext clientSessionCtx) {
+                                    UserSessionModel userSession, ClientSessionContext clientSessionCtx)
+    {
         if (!OIDCAttributeMapperHelper.includeInIDToken(mappingModel)) {
             return token;
         }
         UserModel user = userSession.getUser();
-        String localSub = getLocalIdentifierValue(user, mappingModel);
+        String localSub = HmacPairwiseSubMapperHelper.getLocalIdentifierValue(user, mappingModel);
         if (!checkPrerequisites(localSub, mappingModel, token.getEmail())) {
             return token;
         }
@@ -69,12 +70,13 @@ public class HmacPairwiseEmailMapper extends HmacPairwiseSubMapper {
 
     @Override
     public AccessToken transformAccessToken(AccessToken token, ProtocolMapperModel mappingModel,
-            KeycloakSession session, UserSessionModel userSession, ClientSessionContext clientSessionCtx) {
+                                            KeycloakSession session, UserSessionModel userSession, ClientSessionContext clientSessionCtx)
+    {
         if (!OIDCAttributeMapperHelper.includeInAccessToken(mappingModel)) {
             return token;
         }
         UserModel user = userSession.getUser();
-        String localSub = getLocalIdentifierValue(user, mappingModel);
+        String localSub = HmacPairwiseSubMapperHelper.getLocalIdentifierValue(user, mappingModel);
         if (!checkPrerequisites(localSub, mappingModel, token.getEmail())) {
             return token;
         }
@@ -84,14 +86,15 @@ public class HmacPairwiseEmailMapper extends HmacPairwiseSubMapper {
 
     @Override
     public AccessToken transformUserInfoToken(AccessToken token, ProtocolMapperModel mappingModel,
-            KeycloakSession session, UserSessionModel userSession, ClientSessionContext clientSessionCtx) {
+                                              KeycloakSession session, UserSessionModel userSession, ClientSessionContext clientSessionCtx)
+    {
         if (!OIDCAttributeMapperHelper.includeInUserInfo(mappingModel)) {
             return token;
         }
         UserModel user = userSession.getUser();
-        String localSub = getLocalIdentifierValue(user, mappingModel);
+        String localSub = HmacPairwiseSubMapperHelper.getLocalIdentifierValue(user, mappingModel);
         if (!checkPrerequisites(localSub, mappingModel,
-                String.valueOf(token.getOtherClaims().get("email")))) {
+                                String.valueOf(token.getOtherClaims().get("email")))) {
             return token;
         }
         token.getOtherClaims().put("email", generateEmail(mappingModel, localSub, user.getEmail()));
@@ -101,43 +104,45 @@ public class HmacPairwiseEmailMapper extends HmacPairwiseSubMapper {
     /**
      * Checks whether to execute the mapper.
      */
-    private boolean checkPrerequisites(String localSub, ProtocolMapperModel mappingModel, String email) {
+    private boolean checkPrerequisites(String localSub, ProtocolMapperModel mappingModel, String email)
+    {
         if (localSub == null) {
             return false;
         }
-        boolean overrideEmail = Boolean.valueOf(mappingModel.getConfig().get(OVERRIDE_PROP_NAME));
-        if (!overrideEmail && email != null) {
-            return false;
-        }
-        return true;
+        boolean overrideEmail = Boolean.parseBoolean(mappingModel.getConfig().get(OVERRIDE_PROP_NAME));
+        return overrideEmail || email == null;
     }
 
     /**
      * Generate the HMAC identifier like in {@link HmacPairwiseSubMapper} but add an
      * email domain to it.
      */
-    private String generateEmail(ProtocolMapperModel mappingModel, String localSub, String email) {
-        var pseudoEmail = generateIdentifier(mappingModel, getSectorIdentifier(mappingModel), localSub);
+    private String generateEmail(ProtocolMapperModel mappingModel, String localSub, String email)
+    {
+        var pseudoEmail = HmacPairwiseSubMapperHelper.generateIdentifier(mappingModel, localSub);
         var emailDomain = mappingModel.getConfig().get(EMAIL_DOMAIN_PROP_NAME);
         if (StringUtil.isBlank(emailDomain)) {
             if (!ObjectUtil.isBlank(email)) {
                 pseudoEmail += "@" + email.split("@")[1];
             }
-        } else {
+        }
+        else {
             pseudoEmail += "@" + emailDomain;
         }
         return pseudoEmail;
     }
 
     @Override
-    public List<ProviderConfigProperty> getAdditionalConfigProperties() {
+    public List<ProviderConfigProperty> getAdditionalConfigProperties()
+    {
         var configProperties = super.getAdditionalConfigProperties();
         configProperties.add(createEmailDomainConfig());
         configProperties.add(createOverrideConfig());
         return configProperties;
     }
 
-    private ProviderConfigProperty createOverrideConfig() {
+    private ProviderConfigProperty createOverrideConfig()
+    {
         var property = new ProviderConfigProperty();
         property.setName(OVERRIDE_PROP_NAME);
         property.setType(ProviderConfigProperty.BOOLEAN_TYPE);
@@ -147,7 +152,8 @@ public class HmacPairwiseEmailMapper extends HmacPairwiseSubMapper {
         return property;
     }
 
-    private static ProviderConfigProperty createEmailDomainConfig() {
+    private static ProviderConfigProperty createEmailDomainConfig()
+    {
         var property = new ProviderConfigProperty();
         property.setName(EMAIL_DOMAIN_PROP_NAME);
         property.setType(ProviderConfigProperty.STRING_TYPE);